Split out the modifier methods.
This is by far one of the larger traits. I feel that there is some
opportunity to remove aliased methods. addSecond() and addSeconds() etc.
Don't seem to provide a ton of value.
